rewrite index webpage with spa and update readme
