+++ src/assets/trash.png
