+++ src/components/WelcomeItem.vue
@@ -29,7 +29,6 @@ i {
   place-content: center;
   width: 32px;
   height: 32px;
-
   color: var(--color-text);
 }
