--- src/index.css
@@ -1,114 +1,0 @@
-#banner .masthead {
-    height: 20vh;
-    min-height: 230px;
-    background-image: url('imgsrc/banner.png');
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-}
-
-.header_font {
-    font-weight: bold;
-    font-size: 28px
-}
-
-.title_decoration {
-    background: linear-gradient(to right, #233559 0, #233559 10px, #F0F0F0 10px, #F0F0F0 100%);
-    color: black;
-    font-size: 28px;
-    font-weight: bold;
-}
-
-.more {
-    text-decoration: none;
-    color: navy;
-    float: right;
-}
-
-#calendar {
-    margin-left: 3%;
-    font-family: 'Lato', sans-serif;
-    user-select: none;
-}
-
-#calendar_weekdays div {
-    display: inline-block;
-    vertical-align: top;
-}
-
-#calendar_content,
-#calendar_weekdays,
-#calendar_header {
-    position: relative;
-    overflow: hidden;
-    float: left;
-    z-index: 10;
-}
-
-#calendar_weekdays div,
-#calendar_content div {
-    float: left;
-    align-content: center;
-    overflow: hidden;
-    text-align: center;
-    background-color: #FFFFFF;
-    color: #787878;
-}
-
-#calendar_content {
-    -webkit-border-radius: 0px 0px 12px 12px;
-    -moz-border-radius: 0px 0px 12px 12px;
-    border-radius: 0px 0px 12px 12px;
-}
-
-#calendar_content div:hover {
-    background-color: #F8F8F8;
-}
-
-#calendar_content div.blank {
-    background-color: #E8E8E8;
-}
-
-#calendar_header,
-#calendar_content div.today {
-    zoom: 1;
-    filter: alpha(opacity=70);
-    opacity: 0.7;
-}
-
-#calendar_content div.today {
-    color: #FFFFFF;
-}
-
-#calendar_header {
-    display:flex;
-    text-align: center;
-    background-color: #FF6860;
-    padding: 18px 0;
-    border-radius: 12px 12px 0px 0px;
-}
-
-#calendar_header h1 {
-    display:flex;
-    font-size: 1.5em;
-    color: #FFFFFF;
-    float: left;
-    justify-content: center;
-    align-items: center;
-    padding-top: 8px;
-    width: 70%;
-}
-
-span[class^=icon-chevron] {
-    display:flex;
-    justify-content: center;
-    align-items: center;
-    color: #FFFFFF;
-    float: left;
-    font-weight:bold;
-    font-size: 24px;
-    width: 15%;
-    text-align: center;
-    border-radius: 50%;
-    cursor: pointer;
-}
