--- src/index.html
@@ -1,69 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <title>海大資工課程</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
-    <link rel="stylesheet" href="index.css">
-    <link href='https://fonts.googleapis.com/css?family=Lato' rel='stylesheet' type='text/css'>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
-        crossorigin="anonymous"></script>
-    <script src="https://code.jquery.com/jquery-3.6.1.js"
-        integrity="sha256-3zlB5s2uwoUzrXK3BT7AX3FyvojsraNFxCc2vC/7pNI=" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/1.2.1/axios.js"></script>
-</head>
-
-<body>
-
-    <div id="head"></div>
-    <div id="fixpos"></div>
-    <section id="content">
-        <div class="container my-5">
-            <div class="row">
-                <div class="col-md-12">
-                    <div class="row">
-                        <div class="col-md-8">
-                            <!--單列div-->
-                            <div class="col">
-                                <div class="row-md-12">
-                                    <div class="container row-md-6">
-                                        <div class="title_decoration row h-100 px-4 mb-4">
-                                            最新公告
-                                        </div>
-                                        <div id="bulletin"></div>
-                                        <div id="notification">
-                                            <a href="bulletin/bulletin.html" class="more">>&nbsp;更多...</a>
-                                        </div>
-                                    </div>
-                                    <div class="container row-md-6 my-5">
-                                        <div class="title_decoration row h-100 px-4 mb-4">
-                                            重要時程
-                                        </div>
-                                        <div id="calender"></div>
-                                    </div>
-                                </div>
-                            </div>
-                            <!---->
-                        </div>
-                        <div class="col-md-4">
-                            <div class="container">
-                                <div class="title_decoration row h-100 px-4 mb-4">
-                                    熱門課程
-                                </div>
-                                <div id="hot"></div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </section>
-    <script type="module" src="./banner_scroll.js"></script>
-    <script type="module" src="index_vue.js"></script>
-</body>
-
-</html>
