+++ src/views/LoginView.vue
@@ -1,0 +1,6 @@
+<script setup>
+</script>
+
+<template>
+	<h1>hello!</h1>
+</template>
